Fix warning

+++ platforms/shared/desktop/gui.cpp
@@ -107,7 +107,7 @@ bool gui_init(void)
 
     emu_force_rotation(config_video.rotation);
     emu_audio_mute(!config_audio.enable);
-    emu_audio_set_lowpass_cutoff(config_audio.lowpass_cutoff);
+    emu_audio_set_lowpass_cutoff((float)config_audio.lowpass_cutoff);
     for (int i = 0; i < 4; i++)
         emu_audio_set_volume(i, config_audio.volume[i]);
     emu_debug_set_callback(gui_debug_callback);
